Refactor project structure by moving Bin and BinList types to a new bins package and adding empty packages for api, file, and storage
