+++ 3-struct/bins/bins.go
@@ -1,0 +1,25 @@
+package bins
+
+import "time"
+
+type Bin struct {
+    ID        string
+    Private   bool
+    CreatedAt time.Time
+    Name      string
+}
+
+type BinList []Bin
+
+func NewBin(id, name string, private bool) *Bin {
+    return &Bin{
+        ID:        id,
+        Name:      name,
+        Private:   private,
+        CreatedAt: time.Now(),
+    }
+}
+
+func NewBinList() *BinList {
+    return &BinList{}
+}
